requirements-file.txt
streamlit
llama-index
llama-parse
pinecone-client
voyageai
anthropic
scikit-learn
